# ADPCM JS Optimizations