Optimize ArrayBufferStream and ADPCMSoundDecoder to reduce repeated JS operations (like allocations, and calling to browser internals) and reduce memory used. In brief these optimizations focus on when JavaScript allocates memory and copies data.

- Reuse the array buffer in all extracted ArrayBufferStreams. Removes time creating the copied array buffers and saves memory allocations.
- Reuse a Uint8Array among ArrayBufferStream and extracted children. Repeated typed array view constructions have a measurable impact on CPU and memory. Uint8 is by far the most read integer type from ArrayBufferStream.
- Allocate the Audio Buffer before decoding the samples. The number of samples can be deterministically known ahead of decoding. Storing directly to the channel data for the audio buffer removes two copy operations. One from the first array to a typed array and a second from the typed array to the final channel data array. This also saves time by calling getChannelData once.
- Use file scoped STEP_TABLE and INDEX_TABLE objects. Every access to the ADPCMSoundDecoder.STEP_TABLE static getter allocates and initializes a fresh copy of that array.

### Performance

|  Project - System | Before (seconds) | After (seconds) | Improvement |
|  ------ | ------ | ------ | ------ |
|  130041250 - 2017 MBP 15" Chrome | 0.74 | 0.191 | 287.43% |
|  130041250 - 2017 MBP 15" Firefox | 2.446 | 0.168 | 1355.95% |
|  130041250 - 2017 SMS Chromebook Plus | 2.331 | 0.399 | 484.21% |
|  130041250 - 2018 RPi B+ | 40 | 2.597 | 1440.23% |

Safari on the MBP seems to decode ADPCM itself.

### Memory

The RPi B+ as a test case went from needing 74 MB to <1 MB (from one test run).