@mzgoddard 2018-10-23
This will depend on https://github.com/LLK/scratch-audio/pull/106 to get travis's webpack cli to behave.