@kchadha 2018-10-23
@ericrosenbaum @mzgoddard, I restarted the travis builds and they're passing now.